Fixes #145.
