+++ src/Itinero/RouterDbExtensions.cs
@@ -1834,34 +1834,7 @@ namespace Itinero
         /// </summary>
         public static Coordinate LocationOnNetwork(this RouterDb db, uint edgeId, ushort offset)
         {
-            var geometricEdge = db.Network.GeometricGraph.GetEdge(edgeId);
-            var shape = db.Network.GeometricGraph.GetShape(geometricEdge);
-            var length = db.Network.GeometricGraph.Length(geometricEdge);
-            var currentLength = 0.0;
-            var targetLength = length * (offset / (double)ushort.MaxValue);
-            for (var i = 1; i < shape.Count; i++)
-            {
-                var segmentLength = Coordinate.DistanceEstimateInMeter(shape[i - 1], shape[i]);
-                if (segmentLength + currentLength > targetLength)
-                {
-                    var segmentOffsetLength = segmentLength + currentLength - targetLength;
-                    var segmentOffset = 1 - (segmentOffsetLength / segmentLength);
-                    short? elevation = null;
-                    if (shape[i - 1].Elevation.HasValue && 
-                        shape[i].Elevation.HasValue)
-                    {
-                        elevation = (short)(shape[i - 1].Elevation.Value + (segmentOffset * (shape[i].Elevation.Value - shape[i - 1].Elevation.Value)));
-                    }
-                    return new Coordinate()
-                    {
-                        Latitude = (float)(shape[i - 1].Latitude + (segmentOffset * (shape[i].Latitude - shape[i - 1].Latitude))),
-                        Longitude = (float)(shape[i - 1].Longitude + (segmentOffset * (shape[i].Longitude - shape[i - 1].Longitude))),
-                        Elevation = elevation
-                    };
-                }
-                currentLength += segmentLength;
-            }
-            return shape[shape.Count - 1];
+            return db.Network.LocationOnNetwork(edgeId, offset);
         }
     }
 }
